Project/requirements.txt
llama-index
llama-index-llms-cohere
llama-index-embeddings-cohere
llama-index-vector-stores-pinecone
pinecone-client
gradio
python-dotenv